TG-123: Fix corrections requested
