+++ src/Components/Layout/index.jsx
@@ -6,6 +6,9 @@ import Header from 'Components/Shared/Header';
 import Footer from 'Components/Shared/Footer';
 import Sidebar from 'Components/Shared/Sidebar';
 import styles from './layout.module.css';
+import { tokenListener } from 'helper/firebase';
+import { getAuthAdmin } from 'redux/thunks/admin';
+import { getAuthEmployee } from 'redux/thunks/employee';
 
 const Layout = ({ children }) => {
   const showSidebar = useSelector((state) => state.global.showSidebar);
@@ -13,13 +16,25 @@ const Layout = ({ children }) => {
   const employeeRoutes = useRouteMatch('/employee');
   const adminRoutes = useRouteMatch('/admin');
   const dispatch = useDispatch();
+  const token = useSelector((store) => store.auth.authenticated?.token);
+  const role = useSelector((store) => store.auth.authenticated?.role);
 
   const sidebarToggler = () => {
     if (employeeRoutes || adminRoutes) dispatch(toggleSidebar(true));
     else dispatch(toggleSidebar(false));
   };
 
-  useEffect(() => sidebarToggler(), [location.pathname]);
+  useEffect(() => {
+    sidebarToggler();
+    tokenListener();
+  }, [location.pathname]);
+
+  useEffect(() => {
+    if (token) {
+      role === 'EMPLOYEE' && dispatch(getAuthEmployee());
+      role === 'ADMIN' && dispatch(getAuthAdmin());
+    }
+  }, [token]);
 
   return (
     <div className={styles.container}>
